Move some doctests under the test feature
