+++ src/lib.rs
@@ -106,8 +106,11 @@
 //!
 //! An example of subscribing to an intrinsic event notification for every new [`Win32_Process`]
 //! ```edition2018
-//! # fn main() -> wmi::WMIResult<()> {
 //! # use wmi::*;
+//! # #[cfg(not(feature = "test"))]
+//! # fn main() {}
+//! # #[cfg(feature = "test")]
+//! # fn main() -> WMIResult<()>{
 //! # use serde::Deserialize;
 //! # use std::collections::HashMap;
 //! # let wmi_con = WMIConnection::new(COMLibrary::new()?)?;
@@ -133,7 +136,7 @@
 //! filters.insert("TargetInstance".to_owned(), FilterValue::IsA("Win32_Process"));
 //!
 //! let iterator = wmi_con.filtered_notification::<NewProcessEvent>(&filters)?;
-//! # wmi::start_program();
+//! # tests::start_test_program();
 //!
 //! for result in iterator {
 //!     let process = result?.target_instance;
@@ -269,25 +272,3 @@ pub use variant::Variant;
 #[doc = include_str!("../README.md")]
 #[cfg(all(doctest, feature = "chrono"))]
 pub struct ReadmeDoctests;
-
-// Functions for doc tests. `#[cfg(doctest)]` does not work.
-#[doc(hidden)]
-pub fn start_program() {
-    std::process::Command::new("C:\\Windows\\System32\\cmd.exe")
-        .args(["timeout", "1"])
-        .spawn()
-        .expect("failed to run test program");
-}
-
-#[doc(hidden)]
-pub fn ignore_access_denied(result: WMIResult<()>) -> WMIResult<()> {
-    use winapi::{shared::ntdef::HRESULT, um::wbemcli::WBEM_E_ACCESS_DENIED};
-    if let Err(e) = result {
-        if let WMIError::HResultError { hres } = e {
-            if hres != WBEM_E_ACCESS_DENIED as HRESULT {
-                return Err(e);
-            }
-        }
-    }
-    Ok(())
-}
