+++ src/tests.rs
@@ -1,4 +1,4 @@
-use crate::{COMLibrary, WMIConnection};
+use crate::{COMLibrary, WMIConnection, WMIResult, WMIError};
 
 pub mod fixtures {
     use super::*;
@@ -15,4 +15,23 @@ pub mod fixtures {
 
         wmi_con
     }
+}
+
+pub fn start_test_program() {
+    std::process::Command::new("C:\\Windows\\System32\\cmd.exe")
+        .args(["timeout", "1"])
+        .spawn()
+        .expect("failed to run test program");
+}
+
+pub fn ignore_access_denied(result: WMIResult<()>) -> WMIResult<()> {
+    use winapi::{shared::ntdef::HRESULT, um::wbemcli::WBEM_E_ACCESS_DENIED};
+    if let Err(e) = result {
+        if let WMIError::HResultError { hres } = e {
+            if hres != WBEM_E_ACCESS_DENIED as HRESULT {
+                return Err(e);
+            }
+        }
+    }
+    Ok(())
 }
